Extract common logic in String and Any generators
This will avoid existing and future duplication. If, in the future, we
need Integer single values, the generator should probably be a subclass
of SingleValueStruct.
