+++ app/generators/ruby.rb
@@ -5,6 +5,7 @@ require 'active_support/inflector'
 require_relative 'ruby/type_helpers'
 require_relative 'ruby/base'
 require_relative 'ruby/struct'
+require_relative 'ruby/single_value_struct'
 require_relative 'ruby/array'
 require_relative 'ruby/map'
 require_relative 'ruby/object_from_gobl'
